+++ src/main/java/com/xx_dev/apn/socks/common/ForwardRequest.java
@@ -104,7 +104,7 @@ public class ForwardRequest extends ForwardMsg{
     }
 
     public void encodeAsByteBuf(ByteBuf byteBuf) {
-        byteBuf.writeShort(this.type());
+        byteBuf.writeByte(this.type());
         byteBuf.writeShort(this.streamId());
         byteBuf.writeByte(addressType.byteValue());
         switch (addressType) {
